requirements.txt
schedule
fuzzywuzzy
python-Levenshtein
python-telegram-bot
